organize drive commands
